2022/2022-11-14/fractals/requirements.txt
jupyter==1.0.0
numpy==1.23.4
pillow=9.3.0
matplotlib==3.6.2

